requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-password-strength==1.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
EasyProcess==0.3
idna==2.10
password-strength==0.0.3.post2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
PyVirtualDisplay==1.3.2
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
